- Created handleToken util function. - Created User util functions.
