+++ src/utils/user/index.ts
@@ -1,0 +1,1 @@
+export { default as handleToken } from './handle-token';
